Got bibliography and references working!!
What I did:

Used the bibunits package.
Put \begin{bibunit} <main body>
\putbib[references]
\end{bibunit}
\begin{bibunit}
\nocite{*} %forces all
\putbib[bibliography]
\end{bibunit}

+++ Research/TestBranch/make.sh
@@ -10,9 +10,31 @@ rm *.toc
 
 #make
 pdflatex ${file}.tex
-bibtex ${file}
+if [ $? != 0 ]; then
+	echo Error Occured. Exiting
+	exit 1
+fi
+bibtex bu1.aux
+if [ $? != 0 ]; then
+	echo Error Occured. Exiting
+	exit 1
+fi
+
+bibtex bu2.aux
+if [ $? != 0 ]; then
+	echo Error Occured. Exiting
+	exit 1
+fi
 pdflatex ${file}.tex
+if [ $? != 0 ]; then
+	echo Error Occured. Exiting
+	exit 1
+fi
 pdflatex ${file}.tex
+if [ $? != 0 ]; then
+	echo Error Occured. Exiting
+	exit 1
+fi
  
 #clean
 rm *.aux
